+++ Src/Swagger/ValidationProcessor/Extensions/ValidationExtensions.cs
@@ -80,7 +80,7 @@ internal static class ValidationExtensions
             return propertyName;
 
         var segments = propertyName.Split('.');
-        for (int i = 0; i < segments.Length; i++)
+        for (var i = 0; i < segments.Length; i++)
             segments[i] = namingPolicy.ConvertName(segments[i]);
 
         return string.Join(".", segments);